mk: Add option to register external packages
This adds a target to the subdir role to register a package if it is not
part of the l4 source tree (i.e. an external package). Packages such
registered are tracked in the object directory, built when invoking
global make, built as dependency of other packages if there are any and
also allow to trigger the generation of variants for their dependencies.

As such they integrate better with the more advanced features of the
buildsystem.

Once registered they will be part of global make invocations (i.e. from
the main object or source directory) until the source directory goes
away. Once the source directory goes away they no longer be built.
Unless the variable REMOVE_EXTERNALS is set to `n` external packages
will be unregistered and their object directory will be removed as well
once the source directory goes away.

Beware that installed artifacts (i.e. binaries, libraries, pc files)
will still remain and need to be manually cleaned up since the
buildsystem can no longer track them.

Change-Id: I7b026b62cc6e11308c233451cd2d891cff1aae76