tap-wrapper: drop timeout duration when waiting for additional data
If the test itself was finished, then we only wait for additional data
if wait_for_more is set. Currently this is only used for coverage.

Usually the additional data we wait for gets output quite quickly, so
drop to a reasonable low timeout for waiting for that additional data.

Change-Id: I36bfa0c8bcf848a1d830d058f8983c68ec12cccd