mk/install.inc - Don't call INSTALLFILE_* with a literal $<
Recent changes lead to the INSTALLFILE_* variables being called with a
literal "$<" instead of the actual value. This happened because in
local_install_rule the calls to INSTALLFILE_* are evaluated during the
definition, while the $$< it's being called with is evaluated during the
eval call further down the road. This is an issue if the calls to
INSTALLFILE_* need to examine the parameter, as is the case in lib.mk,
where the INSTALLFILE_* variables check for an *.so suffix to determine
whether to link the file or to copy and strip it.

Change-Id: I22febfda638303b1de473851d3794e4f1b0f8245