mk: Add missing parameter to Lua helper functions
The functions for creating Tap output from an external Lua script when
aborting requires a plugin name, which will be used for prefixing the
reasons for aborting.

Several helper functions called the function to abort, but didn't
specify a plugin name, which results in unhandled errors and incorrect
Tap output.

This change adds the required parameters to the helper functions, passes
it to the abort function, and updates the main plugins that use the
helpers.

Change-Id: I5a0e3149049325084fa244f42006d34b0763660e