L4::TapWrapper - Rework wait_for_more mechanism
The wait_for_more mechanism is a grace period to wait for additional
data even if all plugins signalled they are satisfied. Recently we
adjusted the time span for this grace period from the usual timeout down
to a hardcoded 6 seconds. This does not work in all cases, because in
some cases we need tap-wrapper to wait longer. To facilitate this we let
the Plugin define the time span to wait. And because multiple plugins
accessing one time span variable, might cause interference, we let each
plugin set a separate value, of which the maximum is used for the grace
period.

For backwards compatibility we keep the original $wait_for_more variable
and assume a minimum 6 seconds grace period if it is set to 1.

Change-Id: I74a5385a41b875e0f78ba72370e7e544c157df3c