Plugin::External - Do not start external tool twice
In I3563669a62ef7b324a0399585b6804e85dc07a04 we reworked the External
Plugin to forward stderr from the started process into a dedicated
plugin.log, while stdout was still treated as plain TAP stream. For this
we intended to replace the simple open() with a call to open3(), in
order to open stdin/stdout and stderr as pipes. Unfortunately we failed
to remove the original open() call during that rework, which lead to the
external process being started twice. With now two external processes
meddling in the same plugin temp dir, we observed odd errors about
spontaneously created files that were not meant to exist yet. We fix
this by finally removing the original call to open().

Change-Id: Ie0b27c2c22b69509b9c9bd739dce3c6963975db8