Plugin::LLVMCov - Set different grace periods depending on dump mode
The combination of inhibiting exit + a grace period of 6 seconds does
not work too well if the target writes its data into a memory buffer
instead of the serial line. In the former case we actively expect the
simulator to shutdown, which leads to a failure of the whole test if a
plugin is still inhibiting exit. Instead we do not inhibit exit, but use
a high grace period to "wait for more", in order to wait for the
simulator shutting down on its own.

Change-Id: I16d831b9af10177eb072fe090dfdb27da79580b5