[API] Default to POSIX 7th Edition instead of GNU
This makes GNU extensions unavailable per default. Applications that
want to make use of them need to specify -D_GNU_SOURCE (or
-D_BSD_SOURCE) themselves when required. Note that the C library itself
is still built with these extensions enabled, so that programs using
them can be linked against it. However, there is no guarantee that it
implements all possible extensions.

Change-Id: Ia780e30adb6ba78a0f270239f5379ceb6ee0ab42