L4::Image::EFI - Fix type name field in section header
The name field is not an null-terminated string with a fixed size of 8,
but a zero-padded char array of size 8. The difference is that with the
former a name of length 8 is formatted as 7 chars with a null
terminator, while the latter is formatted as 8 chars without a null
terminator.

Change-Id: I3e1ec07bf5737360547ecc6181c41f27d8ad9e48