L4::Image::EFI - Handle VirtualAddress trailing sections
In an L4Re image, the module data should generally be the last section
regarding the VirtualAddress, because the l4image tool must be able to
grow and shrink the module data section as needed, and most section
types cannot simply be moved to a different VirtualAddress. However
sections with debug information pose an exception to this. In order to
prevent such debug sections from corrupting the module data on boot, we
adjust the VirtualAdress of these trailing PE sections as well.

Change-Id: Id488c6ea1d79b2f0d10e08334b18039cbc55792b