L4::Image::EFI - Preserve COFF Symbol table and String table
The PE section header has an upper limit of 8 characters for the section
name. Longer section names are instead stored in the string table.

For the l4image tool to work, it does not need to read the actual
section name from the symbol table, but if it can preserve the symbol
table the created target image will be a bit easier to inspect.

Change-Id: I1ab239ccba8fe39cc5852a2bb9be82839942599e