Makefile: Avoid symlinking on hard-coded TARGET_IMAGE
When overriding the TARGET_IMAGE through the command line on image
creation we must not symlink the target if the value given is not
different when called with multiple parameters (i.e. contains no
provisions to deal with entry names). Otherwise we will overwrite the
image with a symlink to itself.

Also allow to just rename the image basename without overriding the rest
of the target path and name.

When an entry specific TARGET_IMAGE is desired use

make elfimage TARGET_IMAGE='/tmp/image$(if $2,_$2).$(call IMAGE_SUFFIX,$1)'

or, preferably,

make elfimage IMAGES_DIR=/tmp IMAGE_BASENAME=image

to get '/tmp/image_hello.elf' for the hello entry.

When a hardcoded TARGET_IMAGE, irrespective of entry name or image
suffix, is desired use

make elfimage TARGET_IMAGE=/tmp/myimage.elf

Change-Id: I22138df926120af4df7a5af2eae59d3c5f1ca0a6