binary.inc: Only check dependency consistency from global make
We can only guarantee that the global dependency view is regenerated
when we build from the global Makefile. Thus we should only check that
that global view corresponds to the REQUIRES_LIBS when we do a global
build.

This disables the check for all per-directory builds, but conversely
enables it for out-of-tree packages that have been registered to the
L4Re build tree.

This should address some issues users saw with confusing error messages
saying to fix their Control file, even though it was correct.

Change-Id: I04318586ffa790cb5c5e32e52818eef081cdbea9