Makeconf - Fix issue with syncconfig before menuconfig
If any of the Kconfig files was changed, even if running menuconfig and
friends, the build system would run syncconfig first. This happened
because Makeconf included `.config.all` which depends on `auto.conf`.
`auto.conf` depends on all the Kconfig files and is generated through
syncconfig.

To fix this we simply SKIP the update of the auto.conf if other config
goals are requested.

While we're at it move the list of config goals into its own variable

Change-Id: I41baedcd554a9bddec8d1721f08f039f8a4f4932