bid-bender: use --icf=none to disable identical code folding
This linker parameter will usually not have any impact on the generated
binary when comparing with the default setting (--icf=safe) because the
probability for having identical sections in read-only sections is low.

Disable this link-time optimization to be on the safe side and to behave
like GNU ld.

See also

  https://github.com/llvm-mirror/lld/blob/master/ELF/ICF.cpp

Change-Id: Id56bdf9f141b95ce78fa4ae055f51d5b5cac86cb