tool: Also check for rebuild if using qemu-x86-launch
qemu-x86-launch is a quick way to boot L4Re on x86 qemu using the
binary named "bootstrap" directly and passing the bootstrap modules
via the command line of qemu instead of embedding them into the
bootstrap binary. The problem is that this binary might be
intermittently build as the efi variant. Since the build system did
not check if a rebuild is necessary when using qemu-x86-launch, it was
possible to unintentionally boot the efi variant on a non efi qemu. We
seek to remedy that by adding the same check before calling
qemu-x86-launch.

Change-Id: I80aaaf005ccf1d3044a30a6456e74aeb0aa50d5c