remove usage of create_dir functionality
The create_dir was added to have a forkless way of checking for the
existence of a directory to decide whether to create it. This
feature has been broken for a long time, so we might as well just
use mkdir -p everywhere.

The function is kept for backwards compatibility.

Change-Id: Ibf541bd6e6135c3d8055df110f20d3c751033573