Makefile/checkbuild: relax ext-pkg test
In cases where the ext-pkg directory already exists inside the object
dir, for example due to some additional tests external to checkbuild,
do not complain. Only complain if the ext-pkg directory is created or
modified while doing the checkbuild. This is true if the timestamp of the start of the build is older than the directory modification time.

Change-Id: I5e6ec1efcda3c040ef3c088741a5f959570f6837