Makefile - Do POST_IMAGE_CMD in build system context
Since the introduction of the imagebuilder scripts the POST_IMAGE_CMD
was executed outside of the build system because the image builder
script try to avoid using the build system. This meant that the
POST_IMAGE_CMD did not have any access to build system variables.
However when using one of the make *image commands we do already use the
build system, so we can also run the POST_IMAGE_CMD during its
execution.

In order to do this the imagebuilder_goals in Makefile just add the
POST_IMAGE_CMD at the end of their recipe while not forwarding the
POST_IMAGE_CMD environment variable to the imagebuilder scripts.

Change-Id: I3593defdd9118ede2db341bf66b4ab4d8b585ef5