config.mk: Do not rely on auto.conf presence
After a `make cleanfast` the build directory no longer contains a
include directory. Thus we cannot rely on auto.conf to regenerate
.config.all. It must first be generated. Instead of pretending we
generated it in case we have other config goals while we really
did not, just do not add it as a prerequisite.

This is an addendum to I41baedcd554a9bddec8d1721f08f039f8a4f4932.

Change-Id: Ic8975cd5b9d7e8419eda2763a50fccf34a755a55