mk/rules.inc: Allow asm without cpp
When adding an extension to BID_ASM_FILE_EXTENSIONS but not to
BID_ASM_CPP_FILE_EXTENSIONS, such as .s, then we should not use fixdep,
since the compiler will not pass it through the preprocessor and thus
not use a dependency file.

Also generalize that case that is used for ADAC / F as well and apply it
to AS when applicable.

Change-Id: Ie310f070bbff068cbfd9c7bca1c15de69f5c7d6f