ModList: Detect and include program interpreter
For some reason the `libld-l4.so.1` is listed as a `NEEDED` shared
object in the binary, when linked against uclibc.

However, with musl that is not the case. Also checked on my host Linux,
there the program interpreter also isn't specified as a NEEDED shared
object. A solution is to also scan for the interpreter in ModList.pm.

Unfortunately we need readelf for that, since the hex dump output
provided by objdump is hard to parse.

Change-Id: I73eca9ae8b608af2a13b5232d6837c7c46bf7557