doc: Add rudimentary lua doxygen support
We want to use the present documenation system doxygen to also document
lua code. Doxygen does not come with lua support and only understands a
handful of languages. To add support for other languages you can write
an input filter that translates a file into a language that doxygen
understands. This change primarily focusses on lua functions, so it adds
a script lua2dox.pl, that inspects a lua file, extracts special comments
and translates normal lua function headers into C function prototypes.

The script supports lua single line and multi line comments, but only
extracts those that start with an exclamation mark, meaning
documentation needs to be either after "--!" or between "--[[!" and "]]"

Change-Id: Ib2a9888f30fbe159ec05ea583db6b58a098fef94