mk: Add additional verbosity level
Since we generate cmd files to track changes to compiler command line
options, the command line printed has become quite unwieldy and hard to
copy and paste.

We add a new way to specify multiple verbosity levels, where the new V=2
level will output whole command lines, even if unwieldy, and V=1 should
be restricted to usable command lines and infos.

The new Make function to realize that is verbcmd taking three
arguments which are always executed pasted together (without separators)
and which will be printed as follows:

V=0 / unset: Nothing is printed
V=1: only the middle part of the command is printed
V=2: the whole command sequence is printed

This should hopefully improve output and also serve as a gadget to
improve other outputs in the future.

Change-Id: I8256b7cc72adca1c26c406d40b9b2df349e73f8e
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>